+++ src/fastfetch.c
@@ -58,11 +58,11 @@ static inline void printCommandHelp(const char* command)
 {
     if(command == NULL)
         puts(FASTFETCH_DATATEXT_HELP);
-    else if(ffStrEqualsIgnCase(command, "c") || ffStrEqualsIgnCase(command, "color"))
+    else if(ffStrEqualsIgnCase(command, "color"))
         puts(FASTFETCH_DATATEXT_HELP_COLOR);
     else if(ffStrEqualsIgnCase(command, "format"))
         puts(FASTFETCH_DATATEXT_HELP_FORMAT);
-    else if(ffStrEqualsIgnCase(command, "load-config") || ffStrEqualsIgnCase(command, "loadconfig") || ffStrEqualsIgnCase(command, "config"))
+    else if(ffStrEqualsIgnCase(command, "load-config") || ffStrEqualsIgnCase(command, "config"))
         puts(FASTFETCH_DATATEXT_HELP_CONFIG);
     else if(ffStrEqualsIgnCase(command, "os-format"))
     {
@@ -945,7 +945,7 @@ static void parseOption(FFdata* data, const char* key, const char* value)
     //General options//
     ///////////////////
 
-    else if(ffStrEqualsIgnCase(key, "--load-config"))
+    else if(ffStrEqualsIgnCase(key, "--load-config") || ffStrEqualsIgnCase(key, "--config"))
         optionParseConfigFile(data, key, value);
     else if(ffStrEqualsIgnCase(key, "--gen-config"))
         generateConfigFile(false, value);